requirements.txt
google-genai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
networkx>=3.0
numpy>=1.24.0
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=4.0.0
psycopg2-binary>=2.9.0
